fix: fix error in TextInput and move uploader function to multimodalInput
